Refactor filesystem to not be dependent on a server struct
