+++ server/filesystem/stat_darwin.go
@@ -1,4 +1,4 @@
-package server
+package filesystem
 
 import (
 	"syscall"
