+++ server/filesystem/stat_linux.go
@@ -1,4 +1,4 @@
-package server
+package filesystem
 
 import (
 	"syscall"
