+++ server/filesystem/stat_windows.go
@@ -1,4 +1,4 @@
-package server
+package filesystem
 
 import (
 	"time"
